Link map-graph for any website
Create a visual representation of links on a website with a scatter graph through Plotly. 
The script will check f the necessary libraries are installed and then ask through an input prompt for the link to the website you want to map.
The visualization will be generated in HTML format and should open in a web browser.